OT-125 Use correct argument keys
