+++ lib/delta_chat_core.dart
@@ -42,7 +42,7 @@
 
 import 'dart:async';
 
-import 'package:delta_chat_core/src/event.dart';
+import 'package:delta_chat_core/src/types/event.dart';
 import 'package:flutter/services.dart';
 
 export 'src/base.dart';
@@ -51,8 +51,8 @@ export 'src/chatlist.dart';
 export 'src/chatmsg.dart';
 export 'src/contact.dart';
 export 'src/context.dart';
-export 'src/event.dart';
-export 'src/chatsummary.dart';
+export 'package:delta_chat_core/src/types/event.dart';
+export 'package:delta_chat_core/src/types/chatsummary.dart';
 
 class DeltaChatCore {
   static const String channelDeltaChatCore = 'deltaChatCore';
